requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-allauth==0.44.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==7.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-environ==0.4.5
django-js-asset==1.2.2
django-localflavor==3.0.1
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-seed==0.2.2
django-user-visit==0.4.1
djangorestframework==3.12.2
drf-yasg==1.20.0
Faker==6.6.0
flower==0.9.7
gunicorn==20.1.0
html5lib==1.0.1
humanize==3.2.0
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==2.0.3
nulltype==2.3.1
oauthlib==3.1.0
openapi-codec==1.3.2
ory-kratos-client==0.5.5a2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.19
Pillow==8.1.2
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.17
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-stdnum==1.16
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
razorpay==1.2.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
text-unidecode==1.3
tornado==6.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.2.0
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
